Changes to minimize amount of user configuration needed.
If the bro-pkg config file sets 'scriptpath' and 'pluginpath' to match
the bro installation (e.g. $prefix/share/bro/packages and
$prefix/lib/bro/plugins/packages), then:

- `bro packages` loads the scripts that come from packages
- bro automatically will load plugins that come from packages
- broctl automatically loads broctl plugins that come from packages
- broctl will autoload packages if "@load packages" is in local.bro
- `broctl install` should "just work" for cluster configurations

If installing bro-pkg as part of Bro's CMake setup, 'scriptpath' and
'pluginpath' can likely be preconfigured (not implmented yet),
so all that's up to the user to do (if they want) is
"echo @load packages >> local.bro".

If 'scriptpath' and 'pluginpath' are set to live outside the bro
installation, a bro command-line user can use `bro-pkg env` to
correctly set BROPATH and BRO_PLUGIN_PATH.  Or a standalone broctl
user can add `bro-pkg config scriptpath` to SitePolicyPath and
`bro-pkg config pluginpath` to SitePluginPath in broctl.cfg.