GH-22: add a 'suggests' metadata field
This new field can be used in the same way as 'depends', but is used for
packages that are merely complementary and not strictly required for the
suggesting package to function.

A "(suggested)" indicator for install, bundle, and upgrade commands will
now appear during their bro-pkg confirmation prompts.  These commands
also now have a --nosuggestions flag to ignore all suggested packages.

This changes the bropkg.manager.validate_dependencies() API to return
an additional boolean indicating whether a package was included in the
list as a result of suggestions.