Teach bro-pkg to emit an error when operating on dirty git clones
When trying to use the install, test, or bundle commands on local git
clones that have modifications or untracked files, bro-pkg now aborts
with an error message.  This is to help catch situations where one
thinks their uncommitted local changes will be installed by bro-pkg, but
actually will not be unless they are committed in the git repo.