Improve package info canonicalization
When making a Package object derived from the contents of a manifest
file, there's now a way to prevent git url and name canonicalization as
that should have already been done before that data got written to the
manifest.  Canonicalizing that info again after reading from a manifest
could lead to inconsistencies due to changes in file system state.