Fix unbundling packages that contain submodules
It was not working when absent network access.  The reason was that
there could be an attempt to fetch from a remote repo, but then the
reason for that is a mysterious behavior of "git clone <local dir>" to
inherit the "origin" from <local dir> instead of setting it to be <local
dir>.  And it only seems to do that inheritence when the <local dir>
repo has submodules.  No idea why.