GH-79: fix dependency-aware (un)loading logic to ignore "reserved" names
Both `Manager.load_with_dependencies()` and
`Manger.unload_with_unused_dependers()` previously could find a package
depending on "zeek" (or "zkg") and consider that as the name of a
package to (un)load as part of the operation and fail since they're not
real packages that are installed.  Those operations now simply skip over
such reserved names when walking the dependency graph since there's
nothing to do for them.