GH-76: Detect default branch name of packages automatically
Rather than use the hardcoded 'master' branch as the conventional
default, zkg will now attempt to select a package's default branch as
follows:

If there is a remote named 'origin' and it has a HEAD reference,
whatever branch that points to is taken to be the default branch.

If there is not a remote named 'origin' or it does not have a HEAD,
then the default branch is selected in this order: 'main' if it exists,
'master' if it exists, the currently active branch (HEAD) if there is
one, else the commit hash of the current detached head.

Note that zkg still prioritized release version tags, like 'v1.0.0',
over branches and the default branch selection logic only applies
when no such release version tag exists.