Improve the package source refresh test
This test now also verifies the "refresh --push" behavior. To do this
it makes the "remote" package source bare and verifies the log zkg
generates at the debug level, with a bit of additional canonification.

For better output filtering, the diff-remove-abspath canonifier now
recognizes whitespace (it's the same as used by Zeek), and
diff-remove-zkg-version now actually filters the current zkg version,
not just a related token that appeared in one test.