Added resilience when specifying local paths as package git repos
In addition to failing when a given path doesn't exist, the git.Repo()
constructor also fails when a given path exists but is not a git
repository. That scenario caused backtraces with package installs,
testing, and bundling when providing such invalid local paths.

We now check for this case and provide according output. This adds
guards for the install, test, and bundling commands, which are the
entry points for providing local URLs.

This also expands the test case for invalid repos.