Strengthen the installation-staging concept
The package_base variable in metadata files has different meanings for installed
packages vs those that zkg stages internally. In order to find files via that
variable in either situation, zkg needs to abstract the difference between these
two locations. This commit introduces that abstraction. To provide the material
already available in installed packages, we now symlink the installed clones
into the staging area, with packages under test replacing those symlinks.

This also introduces a class for the staging context, with support for
populating the stage and producing the altered environment for packages.