Support testing package dependencies in Manager.test()
This adds a boolean to the function arguments that, when true, will not just
build and stage the requested package's dependencies, but also test them (in
additin to the requested package). The dependencies are processed in bottom-up
order through the dependency tree, i.e. dependencies are always processed before
any dependers. The flag defaults to False, preserving the old behavior.

This reduces the runtime for package testing from quadratic in the number of
packages in the dependency tree to linear, because previously each requested
package and their dependencies were tested independently, each time building and
staging all dependencies: if a requested package A depends on B and B on C,
we would previously call Manager.test(A), which built/staged A/B/C, then
Manager.test(B), which staged B/C, and finally Manager.test(C). We now only call
Manager.test(A), which as before builds B and C but now can test them, too.