Switch to the sphinx-bundled version of the Napoleon add-on
The Sphinx Napoleon add-on is hitting trouble with Python 3.10:

https://github.com/sphinx-contrib/napoleon/pull/36

The fix is not yet released, meaning the only way to get the zkg docs to build
atm is to install the add-on from their master branch.

Per https://pypi.org/project/sphinxcontrib-napoleon/, Napoleon is bundled with
Sphinx 1.3 (released March 2015), so this switches to the included version.