Make Zeek-provided CMake modules available to packages.
With this patch we make Zeek-provided CMake modules discoverable by
CMake invocations in package script instructions, specifically we append
the directory exposed by `zeek-config --cmake_dir` in the environment
variable `CMAKE_PREFIX_PATH`. If the environment variable is already set
we append to it, so any existing settings take precedence. If script
instructions modify this variable they will find the same CMake modules
if they prepend to the environment variable.

We inject the environment variable by setting it for the whole `zkg`
process at a pretty high level so that it is inherited by any forked
processes like e.g., shells spawned for package script instructions.
Ideally we would use a more general mechanism to set up the environment
for script instructions so that we can use an identical mechanism for
e.g., `PATH`, `CMAKE_PREFIX_PATH` and any future additions.

We also add persistence for this setting by adding it to the zkg
`config` file under `paths.cmake_dir`. Before this patch all variables
in the `paths` section referred to target paths for user settings or
packages (state, script, plugin directory), so this might not be the
optimal ultimate format to store this information.