Reduce to 2 CPUs per task for CI tasks
Specifying 8 CPUs and two tasks immediately allocates as many resources
as are available for free for a single user on Cirrus. Depending on what
else a user is using Cirrus for, execution of zkg tasks may be delayed
for significant amounts of time.

The test-suite does not run such a long duration that the 4x reduction
in CPUs should have a practical impact.