Change --force to fail on failing tests
This is a user-visible change in behavior, but to something sane, so
hopefully that's acceptable.

Currently, `--force` is oftentimes used in scenarios where it should mean
"non-interactive". At the same time, `--force` also implies to ignore
failing test commands. This is not reasonable behavior, particularly as
`--skiptests` is available. This patch changes the behavior and causes
install/upgrade to exit with a failure code when tests fail even with
`--force` provided.

If someone intentionally wants to install a package with broken tests,
they can/should use `--skiptests` explicitly instead. That is less
surprising and stands out more.

Fixes #137