bundle/unbundle: Rework built-in dependency check
Extend unbundle to validate dependencies of contained packages and log a
warning if any are not fulfilled or fail to validate. This can happen when
a bundle is created with --nodeps or the target system has different
built-in packages as compared to the source system.

Built-in packages are included in a bundle's manifest.txt file in a new
[meta] section.