Use default Python version for GH pre-commit workflow
I am unsure why we pinned this to the concrete version in the initial
commit. Looking at it now pinning a linting workflow to e.g., our
minimal Python version seems to make no sense as it does not actually
run any code. Additionally it introduces a requirement of all pre-commit
hooks working with our lowest possible version; this is not the case for
e.g., `pyupgrade` which requires python-3.8 now.