Output most detailed error messages for each version tested
When zkq was running tests for each version, if a version failed, it was
printing an error without context of the test dir and clone dir, and
continuing.

It looks however that the code was meant to not continue after that
first error check, but after the subsequent error, that sets all passed
and prints the more detailed message.

I have changed the behavior to not continue at the first check, and
added
the version under test to the more detailed message.