Use newer tarfile to handle bundles more safely
Use the new `filter` option in tarfile's `extractall()` to mitigate more
issues that could arise from loading a malicious tar file.

Try to normalize permissions when creating the tar and only pay
attention to whether the user `x` bit is set (permissions should be
`0o755` if executable and `0o644` otherwise).

Omit the original username in the tarball and replace with root:root.

When unpacking, the same rules are applied on permission mode, and the
user is left as the user executing zkg.

Use [tarfile.py](https://github.com/python/cpython/raw/v3.8.19/Lib/tarfile.py)
from python 3.8.19 and rename to tarfile_fallback.py so that we can use
the newer features. Once our minimum python versions are greater than
3.12, 3.11.4, 3.10.2, 3.9.17, or 3.8.17, we can safely remove this file
as well as the conditional import in _util.py.