Attempt to make install have similar behavior
Set the umask during `_stage()` in the install process to hopefully give
similar results to bundling and unbundling. Needs btests to verify. We
may need a custom copy function if this doesn't result in the same
behavior as bundling/unbundling.

Looking at this, it might be better to make changes to Manager._stage()
to enforce more consistent permissions when copying.